Rename custom amount dialog to custom amount in file names
