+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/CustomAmountTypeBottomSheet.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.woocommerce.android.R
-import com.woocommerce.android.ui.payments.customamounts.CustomAmountsDialogViewModel.CustomAmountType
+import com.woocommerce.android.ui.payments.customamounts.CustomAmountsViewModel.CustomAmountType
 
 @Composable
 fun CustomAmountTypeBottomSheet(currency: String, onClick: (CustomAmountType) -> Unit) {
